Move core ProcessStatus classes to records.process package
