+++ java/compaction-status-store/src/main/java/sleeper/compaction/status/job/DynamoDBCompactionJobStatusFormat.java
@@ -24,10 +24,10 @@ import sleeper.compaction.job.status.CompactionJobStatus;
 import sleeper.compaction.job.status.CompactionJobStatusesBuilder;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.status.ProcessFinishedStatus;
-import sleeper.core.status.ProcessStartedStatus;
-import sleeper.core.status.ProcessStatusUpdate;
-import sleeper.core.status.ProcessStatusUpdateRecord;
+import sleeper.core.record.process.status.ProcessFinishedStatus;
+import sleeper.core.record.process.status.ProcessStartedStatus;
+import sleeper.core.record.process.status.ProcessStatusUpdate;
+import sleeper.core.record.process.status.ProcessStatusUpdateRecord;
 import sleeper.dynamodb.tools.DynamoDBRecordBuilder;
 
 import java.time.Instant;
