+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/job/StoreCompactionJobUpdatesIT.java
@@ -23,9 +23,9 @@ import sleeper.compaction.status.testutils.DynamoDBCompactionJobStatusStoreTestB
 import sleeper.core.partition.Partition;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.status.ProcessFinishedStatus;
-import sleeper.core.status.ProcessRun;
-import sleeper.core.status.ProcessStartedStatus;
+import sleeper.core.record.process.status.ProcessFinishedStatus;
+import sleeper.core.record.process.status.ProcessRun;
+import sleeper.core.record.process.status.ProcessStartedStatus;
 import sleeper.statestore.FileInfoFactory;
 
 import java.time.Instant;
