+++ java/core/src/main/java/sleeper/core/record/process/status/ProcessFinishedStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.status;
+package sleeper.core.record.process.status;
 
 import sleeper.core.record.process.RecordsProcessedSummary;
