+++ java/core/src/main/java/sleeper/core/record/process/status/ProcessRun.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.core.status;
+package sleeper.core.record.process.status;
 
 import sleeper.core.record.process.RecordsProcessedSummary;
